requirements.txt
apache-airflow
psycopg2-binary
pandas
numpy
python-dotenv